config: PropertyResolver
Further decouple Ribbon's config from a property source by introducing using a PropertyResolver abstraction.
